Got Unity to crash by removing one comma. Making progress now!
This is the point where I decide that this whole approach is crap and start over. The resolution to these compilation, stability, and potentially performance issues is to NEVER convert a string to a Note object. Instead, we'll just pass the Note around directly to dedicated message handling functions (which should be much more efficient).

If we ever need to call On… functions directly, we'll dispatch on the GameObject itself, just for the purpose of the GameObject's use. This may amount to most of the classes doing this, but at least we're not doing any QuackInvoke()s, System.Type.GetType() lookups, and hopefully not many casts, if any.